refactor: Splitting up Component into multiple files that each are only added to the project if used in the project somewhere.
