+++ assets/styles/util/Fill.ts
@@ -1,0 +1,90 @@
+import { exists, ISketchError, isSketchError } from './lang'
+
+const reg = /^#([0-9a-f]{2})([0-9a-f]{2})([0-9a-f]{2})([0-9a-f]{2})?/ig
+const hex = (c: number): string => (c >= 0x10) ? c.toString(16) : `0${c.toString(16)}`
+
+export class RGBA {
+  r: number
+  g: number
+  b: number
+  a: number
+  constructor (string: string) {
+    reg.lastIndex = 0
+    const parts = reg.exec(string)
+    this.r = exists(parts) ? parseInt(parts[1], 16) : 255
+    this.g = exists(parts) ? parseInt(parts[2], 16) : 255
+    this.b = exists(parts) ? parseInt(parts[3], 16) : 255
+    this.a = exists(parts) && exists(parts[4]) ? parseInt(parts[4], 16) : 255
+    this.avg = this.avg.bind(this)
+  }
+
+  toString (): string {
+    return `#${hex(this.r | 0)}${hex(this.g | 0)}${hex(this.b | 0)}${hex(this.a | 0)}`
+  }
+
+  avg (other: RGBA): this {
+    this.r = (this.r + other.r) / 2
+    this.g = (this.g + other.g) / 2
+    this.b = (this.b + other.b) / 2
+    this.a = (this.a + other.a) / 2
+    return this
+  }
+}
+
+export interface IStop {
+  color: string
+  position: number
+}
+
+export enum GradientType {
+  linear = 'linear',
+  radial = 'radial',
+  angular = 'angular'
+}
+
+export interface IGradient {
+  gradient: {
+    type: GradientType
+    stops: IStop[]
+    from: {
+      x: number
+      y: number
+    }
+    to: {
+      x: number
+      y: number
+    }
+  }
+}
+
+function calcAvgColor (stops: IStop[]): string {
+  let rgba: RGBA
+  for (const stop of stops) {
+    const col = new RGBA(stop.color)
+    if (rgba === undefined) {
+      rgba = col
+    } else {
+      rgba = rgba.avg(col)
+    }
+  }
+  return rgba.toString()
+}
+
+export type TFillData = string | IGradient | ISketchError
+
+export const WHITE = '#ffffffff'
+
+export class Fill {
+  data: TFillData
+  color: string
+  constructor (data: TFillData) {
+    this.data = data
+    if (this.data === null || isSketchError(this.data)) {
+      this.color = WHITE
+    } else if (typeof this.data === 'string') {
+      this.color = this.data
+    } else {
+      this.color = calcAvgColor(this.data.gradient.stops)
+    }
+  }
+}
