+++ assets/styles/util/Placement.ts
@@ -1,0 +1,77 @@
+export interface IFrameData {
+  x: number
+  y: number
+  w: number
+  h: number
+}
+
+export interface IStylePlace {
+  left?: number | string
+  top?: number | string
+  width?: number | string
+  height?: number | string
+}
+
+export class Placement {
+  x: number
+  left: number
+  y: number
+  top: number
+  right: number
+  bottom: number
+  width: number
+  height: number
+  centerX: number
+  centerY: number
+
+  constructor ({ x, y, w, h }: IFrameData) {
+    this.x = x
+    this.left = x
+    this.y = y
+    this.top = y
+    this.width = w
+    this.right = this.x + w
+    this.height = h
+    this.bottom = y + h
+    this.centerX = x + w / 2
+    this.centerY = y + h / 2
+  }
+
+  dynWidthStyle (): { left: number, top: number, right: number } {
+    return {
+      left: this.left,
+      top: this.top,
+      right: this.right
+    }
+  }
+
+  size<T extends IStylePlace> (style?: T): T {
+    if (style === undefined || style === null) {
+      style = {} as any
+    }
+    style.width = this.width
+    style.height = this.height
+    return style
+  }
+
+  style<T extends IStylePlace> (style?: T): T {
+    style = this.size(style)
+    style.top = this.top
+    style.left = this.left
+    return style
+  }
+
+  ySpace (other: Placement): number {
+    if (this.y > other.y) {
+      return other.ySpace(this)
+    }
+    return other.top - this.bottom
+  }
+
+  xSpace (other: Placement): number {
+    if (this.x > other.x) {
+      return other.xSpace(this)
+    }
+    return other.x - this.right
+  }
+}
