+++ assets/styles/util/Shadow.ts
@@ -1,0 +1,23 @@
+export interface TShadowData {
+  x: number
+  y: number
+  blur: number
+  spread: number
+  color: string
+}
+
+export class Shadow {
+  x: number
+  y: number
+  blur: number
+  spread: number
+  color: string
+
+  constructor (data: TShadowData) {
+    this.x = data.x
+    this.y = data.y
+    this.blur = data.blur
+    this.spread = data.spread
+    this.color = data.color
+  }
+}
